requirements.txt
python-dotenv
pydantic
requests
openai
google-api-python-client
gspread
oauth2client
